requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pyserial==3.5
pydantic==2.9.2
numpy>=1.24
ultralytics==8.3.27
faster-whisper==1.0.3
piper-tts==1.2.0
ollama==0.3.3
opencv-python-headless>=4.8
